Adding Pole Position for Bones
The only option prior to this commit for poles was to place the pole at the base of a child bone in an IK chain.  Added a new option that places the pole out front of the the kneecap of a leg for example.  To do this specify a distance out front you want it.  To place the pole we take the world space triangle comprising of thigh the knee and the ankle.  If we take a thigh length to knee length adjusted ratio to create a point floating behind the kneecap that when combined with the kneecap location creates a line that can be drawn out of the kneecap to place the pole bone.