Adding Scale for Rigify Fingers
The rigify fingers motion capture scales the finger bone based on the curvature of the fingers.  A capture rig has just finger bones, so we added a scale section that takes in a secondary bone and the base bone of the finger as a source, then scales the finger rig bone based on the dot product of the two source bones.