Adding Quaternion Mode and tweaks to Poles
Quarternion mode keyframes all angels to bones in quats to avoid gimbal lock.  Some animators prefer this although the resulting angles are harder to edit in the graph edittor.  Due to time constraints quaternion mode is does not support twist bones, but these are only used on desitnation rigs with twist bones which are rare.